% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bump_dev_version.R
\name{bump_dev_version}
\alias{bump_dev_version}
\title{Update the version of a package, incrementing dev versions.}
\usage{
bump_dev_version(pkg = ".")
}
\arguments{
\item{pkg}{the path to the package}
}
\value{
the new version
}
\description{
This makes no checks and accepts no responsibility. No backups are made.
}
