% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_backups.R
\name{delete_backups}
\alias{delete_backups}
\title{Delete backup files from the current project}
\usage{
delete_backups(pkg = ".")
}
\arguments{
\item{pkg}{the package to delete files from.}
}
\value{
nothing
}
\description{
Delete backup files from the current project
}
