% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{Document package and non-package projects}
\usage{
document(project.dir = ".", roclets = NULL, quiet = FALSE)
}
\arguments{
\item{project.dir}{the directory of the current (non package project)}

\item{roclets}{Character vector of roclet names to use with package.
The default, \code{NULL}, uses the roxygen \code{roclets} option,
which defaults to \code{c("collate", "namespace", "rd")}.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}
}
\value{
nothing
}
\description{
\code{devtools::document()} requires a DESCRIPTION file, which in general we don't
include in analysis projects. Sometimes we want to produce development
documentation for functions in analysis projects so we can describe options
and behaviour. \code{pkgtools::document} will also work in a non package project
to produce \code{.Rd} files in the \code{man} directory for any functions declared in
the \code{R} subfolder. It also override the default behaviour of \verb{?} to return
the development documentation of matching functions in the current project.
}
\concept{renv}
