% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-stack-tools.R
\name{.create_skeleton}
\alias{.create_skeleton}
\title{Create a skeleton based on a template}
\usage{
.create_skeleton(
  .template = NULL,
  ...,
  .insert_before = "# insert before ----"
)
}
\arguments{
\item{.template}{the whisker template as a character vector}

\item{...}{additional named parameters supplied to the whisker template}

\item{.insert_before}{a marker for where in the file to insert a block if no
matching start and end markers are found.}
}
\value{
A character vector of the file skeleton. does not write the file
}
\description{
Create a skeleton based on a template
}
\keyword{internal}
