% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{.data_description}
\alias{.data_description}
\title{{{title}}}
\format{
An object of class \code{character} of length 1.
}
\source{
<{{.}}>
}
\usage{
.data_description
}
\description{
{{&.}}
}
\details{
{{&.}}
}
\keyword{datasets}
