% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-codegen.R
\name{.delete_fenced_block}
\alias{.delete_fenced_block}
\title{Deletes content between markers in an input}
\usage{
.delete_fenced_block(
  .input,
  name = "block",
  ...,
  .start_glue = "# {name} start ----",
  .end_glue = "# {name} end ----"
)
}
\arguments{
\item{.input}{the input as a character vector}

\item{name}{a name for the block}

\item{...}{additional named parameters supplied to the whisker template}

\item{.start_glue}{a glue specification for the marker of the start of the
block to be replaced. May use any of the whisker variables e.g. \code{"# {name} start ----"}. This can also be given as a numeric index, were particularly
0 is the beginning of the input and -1 is the end. If numeric it is start
inclusive.}

\item{.end_glue}{a glue specification for the marker of the end of the
block to be replaced. May use any of the whisker variables e.g.
\code{"# {name} end ----"} or numeric as above. If numeric it is end inclusive.}
}
\value{
a character vector with the deleted blocks
}
\description{
This function deletes anything between matching markers in an input vector
including the markers themselves. This will match multiple markers if present
and paired
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


tmp = .delete_fenced_block(letters,.start_glue = letters[4], .end_glue = letters[10])
testthat::expect_true(!any(letters[4:10] %in% tmp))

# multiple copies of fenced block are removed
# fenced block to end of file is removed
testthat::expect_equal(
  .delete_fenced_block(rep(letters,3),.start_glue = "b", .end_glue = "z"),
  rep("a",3)
)

# fenced block at beginning is removed:
testthat::expect_equal(
  .delete_fenced_block(letters,.start_glue = "a", .end_glue = "y"),
  "z"
)

# fenced block
testthat::expect_equal(
  .delete_fenced_block(letters,.start_glue = "a", .end_glue = "z"),
  character()
)

testthat::expect_equal(
  .delete_fenced_block(letters,.start_glue = 2, .end_glue = -2),
  c("a","z")
)
}\if{html}{\out{</div>}}
}

