% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-codegen.R
\name{.glue_recover_pieces}
\alias{.glue_recover_pieces}
\title{Extract data from the output of a glue call}
\usage{
.glue_recover_pieces(glue_spec, output)
}
\arguments{
\item{glue_spec}{a glue spec with braces determining named substitutes}

\item{output}{the glue output}
}
\value{
a dataframe with a column per piece which could have been used in
\code{glue::glue_data(.x,glue_spec)} apart from leading white space.
}
\description{
This function allow for the reverse of the \code{glue::glue_data()} operation
and regenerates the data as a dataframe. It won't work for very complex glue
specifications and will only return strings.
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

glue_spec = "# {name}, {value1} and {value2} ----"
output = rep("  # one, two and three ----",5)
tmp = .glue_recover_pieces(glue_spec,output)

testthat::expect_equal(nrow(tmp), 5)
testthat::expect_equal(trimws(output), glue::glue_data(tmp,glue_spec))

}\if{html}{\out{</div>}}
}

