% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-stack-tools.R
\name{.lax_as_function}
\alias{.lax_as_function}
\title{Permissive as_function replacement.}
\usage{
.lax_as_function(fn)
}
\arguments{
\item{fn}{something to interpret as a function or a primitive}
}
\value{
the function or primitive
}
\description{
like \code{rlang::as_function} but only interprets functions or formulae
and ignores primitives and characters.
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{


f1 = .lax_as_function(tolower)
f2 = .lax_as_function(~ tolower(.x))
f3 = .lax_as_function("hello world")

testthat::expect_equal(f1, tolower)
testthat::expect_equal(f2("HELLO WORLD"), f1("HELLO WORLD"))
testthat::expect_equal(f3, f1("HELLO WORLD"))

}\if{html}{\out{</div>}}
}

