% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-df-list-df.R
\name{.list_of_lists_to_df}
\alias{.list_of_lists_to_df}
\title{Convert a multilevel list to a nested dataframe}
\usage{
.list_of_lists_to_df(list)
}
\arguments{
\item{list}{a multilevel list}
}
\value{
a dataframe with each sublist nested as a dataframe
}
\description{
Convert a multilevel list to a nested dataframe
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{
iris_list = .df_to_list_of_lists(iris)
iris2 = .list_of_lists_to_df(iris_list)

testthat::expect_equal(datasets::iris, as.data.frame(iris2))

mtcars_nest = datasets::mtcars \%>\%
  dplyr::mutate(name = rownames(.)) \%>\%
  tidyr::nest(details = -c(cyl,gear))

mtcars_list = mtcars_nest \%>\% .df_to_list_of_lists()

mtcars_nest2 = mtcars_list \%>\% .list_of_lists_to_df()

testthat::expect_equal(
  mtcars_nest2$details[[2]],
  mtcars_nest$details[[2]]
)

}\if{html}{\out{</div>}}
}

