% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-stack-tools.R
\name{.search_call_stack}
\alias{.search_call_stack}
\title{Search the call stack for a first parameter matching a specified class}
\usage{
.search_call_stack(.class, nframe = sys.nframe() - 1)
}
\arguments{
\item{.class}{the class name to find}

\item{nframe}{where to start the search}
}
\value{
a \code{.class} object
}
\description{
Finds the first instance of a function in the call stack that
was called with a object of \code{.class}. This can be used to find
dataframe parameters in tidy calls
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

h = function() {
 df = .search_call_stack(.class = "data.frame")
 return(nrow(df))
}

g = function() {h()}

f = function(x) {g()}

testthat::expect_equal( f(datasets::iris), nrow(datasets::iris))

}\if{html}{\out{</div>}}
}

