% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-find-replace.R
\name{find_and_replace}
\alias{find_and_replace}
\title{Regex find and replace on a codebase}
\usage{
find_and_replace(
  regex,
  replacement,
  pkg = ".",
  rDirectories = c("data-raw", "R", "vignettes", "tests")
)
}
\arguments{
\item{regex}{a PCRE regex}

\item{replacement}{a replacement}

\item{pkg}{the package to scan}

\item{rDirectories}{the subdirectories (defaults to data-raw, R, vignettes, tests)}
}
\value{
a list of changed files
}
\description{
Find and replace within a package after committing the current contents
}
\concept{edit}
