% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_check.R
\name{fix_check}
\alias{fix_check}
\title{Fix common check errors}
\usage{
fix_check(pkg = ".")
}
\arguments{
\item{pkg}{the package to fix}
}
\value{
nothing
}
\description{
This function runs a set of common package development errors. It will create
a commit before running checks and fixing unqualified functions, excluding
non standard files, creating global variables, fixng utf8 encoding issues
and ensuring all dependencies line up.
}
\concept{usethis}
