% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_dependencies.R
\name{fix_dependencies}
\alias{fix_dependencies}
\title{Fixes dependencies in the namespace file using the output of R CMD check.}
\usage{
fix_dependencies(pkg = ".", check)
}
\arguments{
\item{pkg}{the package to scan}

\item{check}{output of a \code{devtools::check()} command
(will check automatically if not provided)}
}
\value{
nothing
}
\description{
Fixes dependencies in the namespace file using the output of R CMD check.
}
\concept{usethis}
