% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_non_standard_files.R
\name{fix_non_standard_files}
\alias{fix_non_standard_files}
\title{Adds non standard and hidden files to the .\code{Rbuildignore} file}
\usage{
fix_non_standard_files(pkg = ".", check)
}
\arguments{
\item{pkg}{the package location}

\item{check}{the results of a \code{devtools::check}}
}
\value{
nothing
}
\description{
Adds non standard and hidden files to the .\code{Rbuildignore} file
}
