% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_unqualified_fns.R
\name{fix_unqualified_fns}
\alias{fix_unqualified_fns}
\title{Fix unqualified functions in active source pane}
\usage{
fix_unqualified_fns()
}
\value{
nothing - called for side effects
}
\description{
Interactively find and replace unqualified, e.g. \code{mutate(...)} calls with
fully qualified \code{dplyr::mutate(...)} calls. This version operates on the
currently open file in the RStudio and allows for interactive review of the
proposed changes.
}
\concept{usethis}
