% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_unqualified_fns.R
\name{fix_unqualified_fns_bulk}
\alias{fix_unqualified_fns_bulk}
\title{Fix errors introduced in package creation by forgetting to qualify
namespaces.}
\usage{
fix_unqualified_fns_bulk(
  pkg = ".",
  rDirectories = c(here::here("R"), here::here("tests/testthat")),
  prioritise = c("dplyr", "rlang", "stringr", "forcats", "ggplot2", "purrr", "tidyr",
    "readr", "stats", "utils")
)
}
\arguments{
\item{pkg}{the package}

\item{rDirectories}{the locations of the R code to fix (by default R scripts,
and tests, but not vignettes)}

\item{prioritise}{a list of package names to pick from first}
}
\value{
nothing. called for side effects.
}
\description{
This is a code linting function and expected to be called at the console
during package development. It will scan the files in the current project and
replace unqualified references to e.g. \code{mutate} with ones to \code{dplyr::mutate}
etc.
}
