% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_utf8_encoding.R
\name{fix_utf8_encoding}
\alias{fix_utf8_encoding}
\title{Fixes utf8 encoded characters in source files replaincg them with \verb{\\uXXXX}}
\usage{
fix_utf8_encoding(pkg = ".", check)
}
\arguments{
\item{pkg}{the package to scan}

\item{check}{output of a \code{devtools::check()} command ()}
}
\value{
nothing
}
\description{
Fixes utf8 encoded characters in source files replaincg them with \verb{\\uXXXX}
}
