% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{generate_data_description}
\alias{generate_data_description}
\title{Auto generate data documentation}
\usage{
generate_data_description(
  data,
  title = sprintf("The `\%s` dataset", name),
  description = sprintf("`\%s` dataset description", name),
  url = "http://example.com",
  name = NULL
)
}
\arguments{
\item{data}{an arbitrary R object}
}
\value{
the data description block formatted in markdown
}
\description{
Auto generate data documentation
}
\examples{
generate_data_description(
  iris,
  title = "The myiris dataset",
  description = c("line 1 & test","line 2","","para 2", paste(rep("asda ",50),collapse="")),
  url = "http://example.com",
  name = "myiris"
)

nested = ggplot2::diamonds \%>\% tidyr::nest(details = -c(cut,color,clarity))

generate_data_description(
  nested,
  title = "The nested diamonds dataset",
  description = "The ggplot2 diamonds data set nested",
  url = "http://example.com",
  name = "mydiamonds"
)
}
