% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemirror5-diff.R
\name{merge_code}
\alias{merge_code}
\title{Merge Code Versions using Shiny and CodeMirror 5 MergeView}
\usage{
merge_code(
  value,
  new,
  old = NULL,
  lhs = "original",
  rhs = "suggested",
  accept = "Save",
  rhs_accept = "Accept all",
  mode = "r"
)
}
\arguments{
\item{value}{the content for the central panel.}

\item{new}{character vector of code lines. Content of the changed / RHS version.}

\item{old}{character vector of code lines. Content of the old / LHS version.}

\item{lhs}{the title of the LHS}

\item{rhs}{the title of the RHS}

\item{accept}{the button text for the LHS}

\item{rhs_accept}{the button text for the RHS}

\item{mode}{the language mode}
}
\value{
Character string of the merged content, or stops with an error if cancelled.
}
\description{
This function launches a Shiny gadget to visually merge two versions of code/YAML.
}
\examples{

if (FALSE) {

  old = paste0("line ",c(2:4,6:12))
  new = paste0("line ",c(1:50))

  merged_result <- merge_code(old, new)
  cat("Merged Content:\n")
  cat(paste0(merged_result, collapse = "\n"),"\n")

}

}
\concept{edit}
