% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate_to_pins.R
\name{migrate_to_pins}
\alias{migrate_to_pins}
\title{Migrate package data to a github hosted pin board}
\usage{
migrate_to_pins(pkg = ".", board = "data", migrate = NULL)
}
\arguments{
\item{pkg}{the package (defaults to current)}

\item{board}{the board name (this is generally not exposed to user)}

\item{migrate}{either "*" for all data files, a list of names or if null the
user will be asked interactively.}
}
\value{
nothing
}
\description{
This function makes multiple changes to the package. It adds packaged data
files from the \code{data} directory to a \code{pkgdown} github published pin board and
constructs a minimal set of functions with the same names as the data to
access the data from the pinboard. It finds any references to the dataset in
the code directories (\code{data-raw},\code{R},\code{vignettes},\code{test}) and replaces them
with a function call of the same name.
}
\concept{usethis}
