% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtools-styler.R
\name{pkgtools_style_transformer}
\alias{pkgtools_style_transformer}
\title{A code transformer for use with \code{\link[styler:style_text]{styler::style_text()}}}
\usage{
pkgtools_style_transformer(..., width = 70)
}
\arguments{
\item{width}{the max width of the code (excludes indent hence approximate)}
}
\description{
This transformer is adapted from \verb{https://github.com/gadenbuie/grkstyle}
It differs in that it will break apart long lines regardless of their initial
line breaks to try and fit them into a set width. It does not unbreak lines.
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

# style_text('testthat::expect_equal(iris[1, ], structure(list(Sepal.Length = 5.1, Sepal.Width = 3.5, Petal.Length = 1.4, Petal.Width = 0.2, Species = structure(1L, levels = c("setosa", "versicolor", "virginica"), class = "factor")), row.names = 1L, class = "data.frame"))')
# style_text('testthat::expect_equal(iris[1, ], structure(list(Sepal.Length = 5.1, Sepal.Width = 3.5, Petal.Length = 1.4, Petal.Width = 0.2, Species = structure(1L, levels = c("setosa", "versicolor", "virginica"), class = "factor")), row.names = 1L, class = "data.frame"))', width=80)
}\if{html}{\out{</div>}}
}

