% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_param.R
\name{roxy_tag_parse.roxy_tag_dparam}
\alias{roxy_tag_parse.roxy_tag_dparam}
\title{Parser for \verb{@dparam} tags}
\usage{
\method{roxy_tag_parse}{roxy_tag_dparam}(x)
}
\arguments{
\item{x}{A tag}
}
\value{
a \code{roxy_tag} object with the \code{val} field set to the parsed value
}
\description{
Replaces the default behaviour of a \verb{@param} tag and documents default
parameters automatically.
}
\examples{
# This provides support to `roxygen2` and only gets executed in the context
# of `devtools::document()`. There is no interactive use of this function.
}
\concept{roxy}
