% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_unit.R
\name{roxy_tag_parse.roxy_tag_unit}
\alias{roxy_tag_parse.roxy_tag_unit}
\title{Parser for \verb{@unit} tags}
\usage{
\method{roxy_tag_parse}{roxy_tag_unit}(x)
}
\arguments{
\item{x}{A tag}
}
\value{
a \code{roxy_tag} object with the \code{val} field set to the parsed value
}
\description{
The \verb{@unit} tag can be used in \code{roxygen2} documentation
of a function to define unit testing code. This code will be written to a
testthat file.
}
\examples{
# This provides support to `roxygen2` and only gets executed in the context
# of `devtools::document()`. There is no interactive use of this function.
}
\concept{document}
