% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_param.R
\name{roxy_tag_rd.roxy_tag_dparam}
\alias{roxy_tag_rd.roxy_tag_dparam}
\title{Support for \verb{@dparam} tags}
\usage{
\method{roxy_tag_rd}{roxy_tag_dparam}(x, base_path, env)
}
\arguments{
\item{x}{The tag}

\item{base_path}{Path to package root directory.}

\item{env}{Environment in which to evaluate code (if needed)}
}
\value{
an \code{roxygen2::rd_section} (see \code{roxygen2} documentation)
}
\description{
Replaces the default behaviour of a \verb{@param} tag and documents default
parameters automatically.
}
\examples{
# An example function definition:
fn_definition <- "
#' This is a title
#'
#' This is the description.
#'
#' @dparam x this parameter specifies a default
#'
f <- function(x = c(\"hello\",\"world\")) {
  return(\"test\")
}
"

# For this example we manually parse the function specification in `fn_definition`
# creating a .Rd block - normally this is done by `roxygen2` which then
# writes this to an .Rd file. This function is not intended to be used
# outside of a call to `devtools::document`.

tmp = roxygen2::parse_text(fn_definition)
print(tmp[[1]])

}
\concept{roxy}
