% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_unit.R
\name{roxy_tag_rd.roxy_tag_unit}
\alias{roxy_tag_rd.roxy_tag_unit}
\title{Support for \verb{@unit} tags}
\usage{
\method{roxy_tag_rd}{roxy_tag_unit}(x, base_path, env)
}
\arguments{
\item{x}{The tag}

\item{base_path}{Path to package root directory.}

\item{env}{Environment in which to evaluate code (if needed)}
}
\value{
an \code{roxygen2::rd_section} (see \code{roxygen2} documentation)
}
\description{
The \verb{@unit} tag can be used in \code{roxygen2} documentation
of a function to define unit testing code. This code will be written to a
testthat file.
}
\examples{
# An example function definition:
fn_definition <- "
#' This is a title
#'
#' This is the description.
#'
#' @md
#' @unit
#' # unit testing code here:
#' testthat::expect_equal(f(), \"test\")
#' @export
f <- function() {
  return(\"test\")
}
"

# For this example we manually parse the function specification in `fn_definition`
# creating a .Rd block - normally this is done by `roxygen2` which then
# writes this to an .Rd file. This function is not intended to be used
# outside of a call to `devtools::document`.

fn_definition = gsub("@","@",fn_definition)
tmp = roxygen2::parse_text(fn_definition)
print(tmp[[1]])
}
\concept{roxy}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{
# for testing
warning("should be suppressed")
}\if{html}{\out{</div>}}
}

