% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_commented_code.R
\name{run_commented_code}
\alias{run_commented_code}
\title{Execute commented out code}
\usage{
run_commented_code()
}
\value{
nothing
}
\description{
RStudio does this for examples but nothing else this allows singly commented
code or code in other tags e.g. \verb{@unit} to be executed
}
\details{
To enable keyboard shortcuts we need the \code{shrtcts} package and \verb{~/.shrtcts.R}

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' Execute commented code
#'
#' Run multi line code stripping off leading comments
#'
#' @shortcut Ctrl+Shift+P
pkgtools::run_commented_code
}\if{html}{\out{</div>}}

and:

in \verb{~/.Rprofile}

\if{html}{\out{<div class="sourceCode">}}\preformatted{if (interactive() && requireNamespace("shrtcts", quietly = TRUE)) \{
  shrtcts::add_rstudio_shortcuts(set_keyboard_shortcuts = TRUE)
\}
}\if{html}{\out{</div>}}
}
\concept{edit}
