% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_renv_repos.R
\name{set_renv_repos}
\alias{set_renv_repos}
\title{Adds new repositories to the beginning of an \code{renv} lockfile}
\usage{
set_renv_repos(..., .wd = here::here())
}
\arguments{
\item{...}{a named list of repository urls}

\item{.wd}{the working directory (defaults to \code{here::here()})}
}
\value{
nothing
}
\description{
Sets custom repositories (e.g. r-universe repositories) in a \code{renv}
lockfile to override CRAN repositories. This is a persistent change can be
undone by manual editing of the lockfile.
}
