% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_standalone.R
\name{standalone_choose}
\alias{standalone_choose}
\title{modified copy of devtools:::standalone_choose}
\usage{
standalone_choose(names, error_call = rlang::caller_env())
}
\arguments{
\item{names}{list of potential standalones from local or remote}

\item{error_call}{where to raise errors}
}
\value{
a single filename which will match standalone-*.R
}
\description{
modified copy of devtools:::standalone_choose
}
