% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_addins.R
\name{standalone_snapshot}
\alias{standalone_snapshot}
\title{Generate a hash based expectation for a standalone object}
\usage{
standalone_snapshot(obj, .clip = interactive(), .as = deparse(substitute(obj)))
}
\arguments{
\item{obj}{the gold standard to test}

\item{.clip}{copy the result to the clipboard (needs \code{clipr})}

\item{.as}{what name to use for the comparison}
}
\value{
a code snippet that can be pasted into the unit test case
}
\description{
\code{testthat} does not use hash based checking anymore, in favour of snapshots.
These don't work well in tests of standalones because the file will move
projects independently of the test directory. This function lets you quickly
create a check based on a gold standard
}
\concept{edit}
