% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtools-styler.R
\name{style_text}
\alias{style_text}
\title{Style code in \code{pkgtools} style}
\usage{
style_text(text, ..., transformers = pkgtools_style_transformer(...))
}
\arguments{
\item{text}{A character vector with text to style.}

\item{...}{
  Named arguments passed on to \code{\link[styler:style_text]{styler::style_text}}\describe{
    \item{\code{...}}{Arguments passed on to the \code{style} function,
see \code{\link[styler:tidyverse_style]{tidyverse_style()}} for the default argument.}
\item{\code{style}}{A function that creates a style guide to use, by default
\code{\link[styler]{tidyverse_style}}. Not used
further except to construct the argument \code{transformers}. See
\code{\link[styler:style_guides]{style_guides()}} for details.}
\item{\code{include_roxygen_examples}}{Whether or not to style code in roxygen
examples.}
\item{\code{base_indention}}{Integer scalar indicating by how many spaces the whole
output text should be indented. Note that this is not the same as splitting
by line and add a \code{base_indention} spaces before the code in the case
multi-line strings are present. See 'Examples'.}
}
  Named arguments passed on to \code{\link[=pkgtools_style_transformer]{pkgtools_style_transformer}}\describe{
    \item{\code{width}}{the max width of the code (excludes indent hence approximate)}
}}

\item{transformers}{A set of transformer functions. This argument is most
conveniently constructed via the \code{style} argument and \code{...}. See
'Examples'.}
}
\value{
the formatted code
}
\description{
This is adapted from \verb{https://github.com/gadenbuie/grkstyle}
It differs in that it will break apart long lines regardless of their initial
line breaks to try and fit them into a set width. It does not unbreak lines.
}
