% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_addins.R
\name{switch_expect_equals}
\alias{switch_expect_equals}
\title{Switch expression for equality based test case.}
\usage{
switch_expect_equals()
}
\value{
nothing
}
\description{
Switch expression for equality based test case.
}
\concept{edit}
