% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_addins.R
\name{switch_standalone_snapshot}
\alias{switch_standalone_snapshot}
\title{Switch global variable for hash based test case.}
\usage{
switch_standalone_snapshot()
}
\value{
nothing
}
\description{
Switch global variable for hash based test case.
}
\concept{edit}
