% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_standalone.R
\name{sync_standalone_to_master}
\alias{sync_standalone_to_master}
\title{Synchronise master version of standalone file with editor contents}
\usage{
sync_standalone_to_master()
}
\value{
nothing used for side effects
}
\description{
This file syncs the current content of the rstudio editor window with
a local master standalone file, which is assumed to be in a sibling project
directory. This allows for changes to be propagated back to the master when
in active development.
}
