% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_template.R
\name{use_template}
\alias{use_template}
\title{Use a template file}
\usage{
use_template(.template_name = NULL, ..., .overwrite = FALSE, .destfile = NULL)
}
\arguments{
\item{.template_name}{if not given a dialog will be presented with the
options}

\item{...}{supplied as parameters to the template. template placeholders such
as \code{`xxx.#name#_yyy`}, will be substituted with \code{xxx.myname_yyy}
if \code{name="myname"}. The first value of this list will be used for
\code{`##`} placeholders}

\item{.overwrite}{overwrte the destination file if it exists}

\item{.destfile}{if given will write the expanded template here.}
}
\value{
the written file name
}
\description{
Use a template file
}
\concept{usethis}
