% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_unit.R
\name{use_unit_test}
\alias{use_unit_test}
\title{Use '@unit' unit tests in your package}
\usage{
use_unit_test(path = ".")
}
\arguments{
\item{path}{character path to the package}
}
\value{
Nothing: called for file system side effects.
}
\description{
This functions amends DESCRIPTION to set up pkgtools usage.
Call it once when you're setting a new package.
}
\concept{usethis}
