% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-df-list-df.R
\name{.df_to_list_of_lists}
\alias{.df_to_list_of_lists}
\title{Convert a nested dataframe to a multilevel list}
\usage{
.df_to_list_of_lists(df)
}
\arguments{
\item{df}{a nested dataframe}
}
\value{
a list of lists
}
\description{
Convert a nested dataframe to a multilevel list
}
\keyword{internal}

\section{Unit tests}{
\code{
\preformatted{

iris_list = .df_to_list_of_lists(iris)

testthat::expect_equal(
  iris_list[[1]]$Species,
  iris$Species[[1]]
)

mtcars_nest = mtcars \%>\%
  dplyr::mutate(name = rownames(.)) \%>\%
  tidyr::nest(details = -c(cyl,gear))

mtcars_list = mtcars_nest \%>\% .df_to_list_of_lists()

testthat::expect_equal(
  mtcars_list[[1]]$details[[1]]$name,
  mtcars_nest$details[[1]]$name[[1]]
)

}
}
}

