% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-codegen.R
\name{.glue_detect_output}
\alias{.glue_detect_output}
\title{Detect glue outputs}
\usage{
.glue_detect_output(glue_spec, content)
}
\arguments{
\item{glue_spec}{a glue spec with braces determining named substitutes}

\item{content}{a character vector}
}
\value{
a logical vector with true if the character could have been created
by the glue spec.
}
\description{
This allows you to filter for lines that match those that could have been
generated with a given glue spec
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

content = c("sdlfksl","  # X, Y and Z ----","sdfskjfsdlk asdlkfjsdlkj")
testthat::expect_equal(
  .glue_detect_output("# {name}, {value1} and {value2} ----", content),
  c(FALSE,TRUE,FALSE)
)
}\if{html}{\out{</div>}}
}

