% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_dependencies.R
\name{fix_dependencies}
\alias{fix_dependencies}
\title{Fixes dependencies in the namespace file using the output of R CMD check.}
\usage{
fix_dependencies(pkg = ".", check = qcheck(quiet = TRUE))
}
\arguments{
\item{pkg}{the package to scan}

\item{check}{output of a \code{devtools::check()} command ()}
}
\value{
a list of the
}
\description{
Fixes dependencies in the namespace file using the output of R CMD check.
}
