% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_utf8_encoding.R
\name{fix_utf8_encoding}
\alias{fix_utf8_encoding}
\title{Fixes dependencies in the namespace file using the output of R CMD check.}
\usage{
fix_utf8_encoding(pkg = ".", check = qcheck(quiet = TRUE), dry_run = FALSE)
}
\arguments{
\item{pkg}{the package to scan}

\item{check}{output of a \code{devtools::check()} command ()}

\item{dry_run}{test changes without breaking originals.}
}
\value{
a list of the
}
\description{
Fixes dependencies in the namespace file using the output of R CMD check.
}
