% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_addins.R
\name{run_commented_code}
\alias{run_commented_code}
\title{Execute commented out code}
\usage{
run_commented_code()
}
\value{
nothing
}
\description{
RStudio does this for examples but nothing else this allows singly commented
code or code in other tags e.g. \verb{@unit} to be executed
}
\details{
This is installed as an RStudio add in and can be given a keyboard shortcut
through \code{Tools} > \verb{Modify Keyboard Shortcuts...}
}
\concept{edit}
