% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-git.R
\name{undo}
\alias{undo}
\title{Undo a bulk \code{pkgtools} operation}
\usage{
undo(pkg = ".")
}
\arguments{
\item{pkg}{the package}
}
\value{
nothing
}
\description{
Reverts to last state committed by pkgtools, and stashes any changes.
}
