% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstable.R
\name{unstable}
\alias{unstable}
\title{Reload a set of packages that are on the local machine}
\usage{
unstable(path = ".", ..., force = TRUE, upgrade = "never", quiet = TRUE)
}
\arguments{
\item{path}{the package local development repository path. This assumes you
have all your other packages code in a sibling directory}

\item{...}{
  Arguments passed on to \code{\link[remotes:install_local]{remotes::install_local}}
  \describe{
    \item{\code{subdir}}{subdirectory within url bundle that contains the R package.}
    \item{\code{dependencies}}{Which dependencies do you want to check?
Can be a character vector (selecting from "Depends", "Imports",
"LinkingTo", "Suggests", or "Enhances"), or a logical vector.

\code{TRUE} is shorthand for "Depends", "Imports", "LinkingTo" and
"Suggests". \code{NA} is shorthand for "Depends", "Imports" and "LinkingTo"
and is the default. \code{FALSE} is shorthand for no dependencies (i.e.
just check this package, not its dependencies).

The value "soft" means the same as \code{TRUE}, "hard" means the same as \code{NA}.

You can also specify dependencies from one or more additional fields,
common ones include:
\itemize{
\item Config/Needs/website - for dependencies used in building the pkgdown site.
\item Config/Needs/coverage for dependencies used in calculating test coverage.
}}
    \item{\code{build}}{If \code{TRUE} build the package before installing.}
    \item{\code{build_opts}}{Options to pass to \verb{R CMD build}, only used when \code{build} is \code{TRUE}.}
    \item{\code{build_manual}}{If \code{FALSE}, don't build PDF manual ('--no-manual').}
    \item{\code{build_vignettes}}{If \code{FALSE}, don't build package vignettes ('--no-build-vignettes').}
    \item{\code{repos}}{A character vector giving repositories to use.}
    \item{\code{type}}{Type of package to \code{update}.}
  }}

\item{force}{Force installation, even if the remote state has not changed
since the previous install.}

\item{upgrade}{Should package dependencies be upgraded? One of "default", "ask", "always", or "never". "default"
respects the value of the \code{R_REMOTES_UPGRADE} environment variable if set,
and falls back to "ask" if unset. "ask" prompts the user for which out of
date packages to upgrade. For non-interactive sessions "ask" is equivalent
to "always". \code{TRUE} and \code{FALSE} are also accepted and correspond to
"always" and "never" respectively.}

\item{quiet}{If \code{TRUE}, suppress output.}
}
\value{
nothing
}
\description{
Vignette building uses a new session. Any changes in current project or
dependent projects are not tested unless the packages are all installed using
\code{devtools::install_local(...)}. This causes problems when developing multiple
packages in parallel.
}
\details{
This function scans the current package and first order dependencies, looking
for local development directories for any packages imported. Looks for
changes in files in local development directories of package and first order
dependencies versus files currently installed in r library.

Any recent file change in development directories triggers a dev version bump
and local package installation. After a call to \code{unstable()} any dependencies
in your local dev environment are up to date.
}
