% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_has_changed.R
\name{what_has_changed}
\alias{what_has_changed}
\title{Compare content of editor with last saved version}
\usage{
what_has_changed()
}
\value{
nothing
}
\description{
I frequently end up getting an unsaved change in RStudio conflicitng with
an update on disk. Often as a result of a global find and replace. This helps
determine which version to keep.
}
